Move ZK implementation back into interface file
